requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
Django==1.10.3
django-appconf==1.0.2
django-bootstrap-form==3.2.1
django-bootstrap-toolkit==2.15.0
django-braces==1.9.0
django-cors-headers==1.1.0
django-debug-toolbar==1.6
django-forms-bootstrap==3.0.1
django-oauth-toolkit==0.10.0
django-social-auth==0.7.28
django-user-accounts==2.0.0
django-uwsgi==0.1.3
django-widget-tweaks==1.4.1
djangorestframework==3.4.4
docopt==0.4.0
facebook-sdk==2.0.0
httplib2==0.9.2
kombu==3.0.35
mailchimp==2.0.9
mailchimp3==1.0.26
mysqlclient==1.3.7
oauth2==1.9.0.post1
oauthlib==1.0.3
Pillow==3.4.2
python-memcached==1.58
python-openid==2.2.5
pytz==2016.7
redis==2.10.5
requests==2.11.0
requests-oauthlib==0.6.2
six==1.10.0
sqlparse==0.2.2
sugarcrm==0.1.2
uWSGI==2.0.13.1
